Change namespace to io.github.sinuscosinustan
The old namespace `me.tomsdevsn` got banned from ossrh because
the domain expired and was risked to be taken over by a malicious
actor.

Even after the efforts to gain back the access to deploy to
`me.tomsdevsn`, I was not able to deploy new versions to this namespace.

After some thoughts it only made sense to change the whole namespace
to a more stable one (in this case to `io.github.sinuscosinustan`),
which will very likely not change for the next years.

Signed-off-by: Tan Siewert <tan@siewert.io>